requirements.txt
Django
djangorestframework
djangorestframework-simplejwt==5.2.2
psycopg2
djangorestframework-jwt
django-tenant-schemas
psycopg2-binary
django-tenants
django-jazzmin
django-cors-headers
gunicorn
marshmallow

